Refactor the hierarchy of NavigationMode classes.
Shared code extracted to the base `abstract class NavigationMode`.

This higlights differences between provided navigation modes.
